30-07-2024 Add new block 'products' for page 'Category'.

+++ src/main/java/com/techmatrix18/repository/ProductRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface ProductRepository extends CrudRepository<Product, Long> {
     Product getById(Long id);
     List<Product> findByTitle(String title);
-    List<Product> findByCategoryId(Integer categoryId);
+    List<Product> findByCategoryId(Long categoryId);
 
     List<Product> findByWeight(Integer distance);
     //List<Product> findByWeightLess(Integer distance);
